requirements.txt
setuptools
ollama
gitpython
argparse
g4f[all]
rich
requests
openai
groq
google-generativeai